First draft of component manager
